--- src/ai/requests/tests/apple/message.json
@@ -1,14 +1,0 @@
-{
-  "id": "sim",
-  "role": "assistant",
-  "parts": [
-    {
-      "type": "reasoning",
-      "text": "\nOkay, the user is asking for the weather forecast this week. Let me check what I need to do.\n\nFirst, I remember that the user hasn't provided their location yet. The instructions say I should ask for the location before using any location-based tools. Since the weather forecast depends on the location, I can't proceed without that information. \n\nI should respond by asking them where they are located. That way, once they provide the city or area, I can use the appropriate tool to get the forecast. I need to make sure I don't mention any tools by name, just ask for the location. \n\nAlso, I need to follow the format: use Markdown, subheaders, bullet points, and emojis if appropriate. Let me structure the response politely and clearly. Make sure to explain why I need the location so they understand it's necessary for the forecast. \n\nDouble-checking the guidelines: don't invent info, be honest if I can't help without the location. Yep, that's covered. Alright, time to put it all together.\n"
-    },
-    {
-      "type": "text",
-      "text": "\n\n🌤️ **Weekly Weather Forecast Request**  \n\nTo provide you with the most accurate forecast, I need to know your **location** (e.g., city or region). Weather varies by area, and real-time data requires this detail to proceed.  \n\nCould you share where you're located? Once I have that, I’ll fetch the latest forecast for you! 🌍✨"
-    }
-  ]
-}
