--- src/ai/requests/tests/banana/message.json
@@ -1,10 +1,0 @@
-{
-  "id": "sim",
-  "role": "assistant",
-  "parts": [
-    {
-      "type": "text",
-      "text": "Hello! 👋 How can I assist you today? Whether you have questions, need help with planning, or just want to chat, I'm here for you! Let me know how I can be helpful. 🌟  \n\n**Need help with:**  \n- Finding information?  \n- Scheduling or reminders?  \n- General advice or ideas?  \n\nJust ask! 😊"
-    }
-  ]
-}
