+++ src/ai/streaming/sse-logs/apple.sse
