+++ src/ai/streaming/sse-logs/banana.sse
