+++ src/ai/streaming/util.ts
